Updating trunk. Currently broken. Errors: - post count in Archive is fucked up - clean installation might still "generate unexpected output" - seems to have crashed my remote server (local seems to not mind though...)
Need to sleep. bbl. :/

Updates:
* Added: support and widget for custom taxonomies. (Thanks: Einat Dagan!)
* Added: configurable open/close links for each tree
* Added: [configurable open/close separator for each tree](http://wordpress.org/support/topic/plugin-wp-dtree-pipe-character-between-open-all-and-close-all?replies=2)
* Fixed: various PHP [errors](http://wordpress.org/support/topic/plugin-wp-dtree-stripos-function-error-followed-by-crash-of-site?replies=1#post-2393280) and warnings.
* [Fixed: feed links being wrong](http://wordpress.org/support/topic/rss-problem-36?replies=2#post-2221275)
* Fixed: initialization routines
* Fixed: excluding categories from the archive tree
* Fixed: excluding posts from category tree
* Fixed: [open/close link separator behavior](http://wordpress.org/support/topic/plugin-wp-dtree-pipe-character-between-open-all-and-close-all?replies=2)
* Fixed: [link children are sorted correctly](http://wordpress.org/support/topic/plugin-wp-dtree-sorting-links-by-id-not-working?replies=1)
* Fixed: Open to selection supports pagination
* Fixed: [Open to selection supports categories again](http://wordpress.org/support/topic/plugin-wp-dtree-how-to-force-open-wdtree?replies=3#post-2523443)


git-svn-id: https://plugins.svn.wordpress.org/wp-dtree-30/trunk@482892 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ wp-dtree-cache.php
@@ -3,33 +3,35 @@
 function wpdt_get_table_name(){
 	global $wpdb; return $wpdb->prefix . "dtree_cache";
 }
+
 function wpdt_install_cache(){	
+	require_once(ABSPATH . 'wp-admin/includes/upgrade.php');
 	global $wpdb;
 	$wpdt_cache = wpdt_get_table_name();
 	wpdt_uninstall_cache();		
 	$charset_collate = '';
 	if(version_compare(mysql_get_server_info(), '4.1.0', '>=')){
 		if(!empty($wpdb->charset)){
-			$charset_collate = "DEFAULT CHARACTER SET $wpdb->charset";
+			$charset_collate = "DEFAULT CHARACTER SET {$wpdb->charset}";
 		}
 		if(!empty($wpdb->collate)){
-			$charset_collate .= " COLLATE $wpdb->collate";
+			$charset_collate .= " COLLATE {$wpdb->collate}";
 		}
 	}			
 	$sql = "CREATE TABLE {$wpdt_cache} (
 	hash BINARY(16) NOT NULL, 
 	content MEDIUMTEXT NOT NULL,				
 	UNIQUE KEY  hash (hash)		
-	) {$charset_collate};";
-	$wpdb->show_errors();
-	require_once(ABSPATH . 'wp-admin/includes/upgrade.php');		
-	dbDelta($wpdb->prepare($sql));					
+	) {$charset_collate};";	
+	dbDelta($sql);					
 }
 
 function wpdt_uninstall_cache(){			
 	global $wpdb;
 	$wpdt_cache = wpdt_get_table_name();	
-	$wpdb->query("DROP TABLE " . $wpdt_cache); 	
+	if($wpdb->get_var("SHOW TABLES LIKE '$wpdt_cache'") == $wpdt_cache) {
+		$wpdb->query("DROP TABLE " . $wpdt_cache); 	
+	}	
 }
 
 /*we no longer have a single monoholotic cache.
@@ -89,13 +91,6 @@ function wpdt_clean_exclusion_list($excluded){
 }
 
 function wpdt_build_exclude_statement($excluded, $field ='ID'){
-	$exclusions = '';	
-	$excluded = preg_split('/[\s,]+/',$excluded);
-	if(count($excluded)){
-		foreach($excluded as $ex){
-			$exclusions .= " AND {$field} != " . intval($ex) . ' ';
-		}
-	}
-	return $exclusions;
+	return ($excluded) ? " AND {$field} NOT IN ($excluded) " : '';
 }
 ?>
